hw2/requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud-storage==1.22.0
google-cloud-bigquery==1.22.0

